Add custom display names for LOINC codes
Let users rename any lab metric (e.g. 4548-8 → "HbA1c %") from Settings →
Sort & Visibility. The override is stored per code in LabDisplayPreferences
(roams via iCloud like the rest of the layout) and applied centrally in
LabMapping.displayName, so it flows to every screen — dashboard, trends,
detail, review, history. It is cosmetic only: the exported CDA keeps the
canonical LOINC names, so renames never leak into saved documents.

Extracted LabSortEditor into its own file to stay under the file-length limit.
